Drop support for sub-second precision timestamps
Convert the batch table columns to integers, and change the
DatabaseStore timestamp conversion methods to only work on full-second /
integer timestamps. This also affects the serialization of “maxlag
exceeded” command records, so those are changed to truncate to the full
second as well; the upgrade maintenance script does not update any
existing JSON serializations of such command records, since they should
be rare (none exist on Toolforge, and I’ve updated the four timestamps
on my local installation manually).

+++ tables.sql
@@ -4,8 +4,8 @@ CREATE TABLE batch (
   batch_local_user_id int unsigned NOT NULL,
   batch_global_user_id int unsigned NOT NULL,
   batch_domain_id int unsigned NOT NULL,
-  batch_created_utc_timestamp double unsigned NOT NULL,
-  batch_last_updated_utc_timestamp double unsigned NOT NULL,
+  batch_created_utc_timestamp int unsigned NOT NULL,
+  batch_last_updated_utc_timestamp int unsigned NOT NULL,
   batch_status int unsigned NOT NULL
 )
 CHARACTER SET = 'utf8mb4'
